refactor(pipeline): move per-turn definition into lib/pipeline/definitions
Pipeline definitions are declarative content that writes exclusively through emitted events, not imperative store+SQLite actions — lib/actions keeps only the trigger (submitTurn). Sets the placement pattern before chapter-close/classifier land, which span too many domains for any lib/actions home. Placement rule recorded in generation-pipeline.md. M2-reconciliation note 3.
