+++ lib/actions/index.ts
@@ -40,7 +40,6 @@ export { ENTRIES_WINDOW_SIZE, readRecentEntries } from './story-entries/recent-w
 export { STORY_ENTRY_REJECTION, type StoryEntryRejectionCode } from './story-entries/register'
 export { redoLastAction, undoLastAction, type UndoResult } from './story-entries/undo'
 export { clearSystemEntry, writeSystemEntry } from './story-entries/system-entry'
-export { PER_TURN_KIND } from './turns/pipeline'
 export { submitTurn, type SubmitTurnMeta } from './turns/submit-turn'
 export type { DbCtx, DeltaSource, MutationResult, PipelineAction } from './types'
 export {
