+++ lib/pipeline/index.ts
@@ -1,5 +1,6 @@
 export { toPipelineError } from './call-error'
 export { definePhase, definePipeline } from './authoring/define'
+export { ensurePerTurnPipelineRegistered, PER_TURN_KIND } from './definitions/per-turn'
 export { __resetRegistry, getPipeline, registerPipeline } from './authoring/registry'
 export { __resetBus, pipelineEventBus } from './runtime/event-bus'
 export {
